+++ consensus/wal.go
@@ -20,17 +20,13 @@ import (
 //--------------------------------------------------------
 // types and functions for savings consensus messages
 
-var (
-	walSeparator = []byte{55, 127, 6, 130} // 0x377f0682 - magic number
-)
-
 type TimedWALMessage struct {
 	Time time.Time  `json:"time"` // for debugging purposes
 	Msg  WALMessage `json:"msg"`
 }
 
 // EndHeightMessage marks the end of the given height inside WAL.
-// @internal used by scripts/cutWALUntil util.
+// @internal used by scripts/wal2json util.
 type EndHeightMessage struct {
 	Height int64 `json:"height"`
 }
@@ -195,7 +191,7 @@ func NewWALEncoder(wr io.Writer) *WALEncoder {
 }
 
 // Encode writes the custom encoding of v to the stream.
-func (enc *WALEncoder) Encode(v interface{}) error {
+func (enc *WALEncoder) Encode(v *TimedWALMessage) error {
 	data := wire.BinaryBytes(v)
 
 	crc := crc32.Checksum(data, crc32c)
@@ -208,11 +204,6 @@ func (enc *WALEncoder) Encode(v interface{}) error {
 	copy(msg[8:], data)
 
 	_, err := enc.wr.Write(msg)
-
-	if err == nil {
-		// TODO [Anton Kaliaev 23 Oct 2017]: remove separator
-		_, err = enc.wr.Write(walSeparator)
-	}
 
 	return err
 }
@@ -278,26 +269,7 @@ func (dec *WALDecoder) Decode() (*TimedWALMessage, error) {
 		return nil, fmt.Errorf("failed to decode data: %v", err)
 	}
 
-	// TODO [Anton Kaliaev 23 Oct 2017]: remove separator
-	if err = readSeparator(dec.rd); err != nil {
-		return nil, err
-	}
-
 	return res, err
-}
-
-// readSeparator reads a separator from r. It returns any error from underlying
-// reader or if it's not a separator.
-func readSeparator(r io.Reader) error {
-	b := make([]byte, len(walSeparator))
-	_, err := r.Read(b)
-	if err != nil {
-		return fmt.Errorf("failed to read separator: %v", err)
-	}
-	if !bytes.Equal(b, walSeparator) {
-		return fmt.Errorf("not a separator: %v", b)
-	}
-	return nil
 }
 
 type nilWAL struct{}